Rename client module to session
